style: 타입의 상관관계에 따른 파일 분리
